Update patch set 1
Patch Set 1:

> If the new transforms wrote to 8-bit buffers (even with a slow copy)

Yes, that's what I was thinking about.
This makes the (implementability) implications of dropping the LBD transforms very limited (compared to fully dropping the LBD coding path).

> the end to end tests [...] IMO should match what the user gets via ./configure. 

Not all end-top-end test are equal ; it depends on what behaviour is being tested.

Take av1_ext_tile_test.cc: this test is end-to-end and will fail if, for example, there's a crash in transform functions or in the rate control.

However, this is not what's being tested here.

Trying to catch other unrelated bugs using this test is dangerous, and relies on the good will of the RD optimizer inside the encoder to choose encoding modes that are "interesting for validation" (actually, the RD optimizer probably does the exact opposite).

My point is the goal of this test isn't to be end-to-end. It simply happens to be the easiest way (but not the simplest) to implement the check about tile mode.

Switching the underlying implementation for such a test, by making it depend on CONFIG_LOWBITPDEPTH, would basically increase its fragility, adding new unwanted reasons for it to fail.

> This is a very good idea, and an argument for keeping a runtime flag to switch between the two paths.

Thanks, here it is: https://aomedia-review.googlesource.com/c/14960/

Please let me know what you think.
James: if you're reading this, I'd be very interested to hear your opinion about this.

> There are also the CONFIG_DAALA_DCT* experiments currently in Gerrit, these are transforms that both require smaller intermediates, and are identical in both bit depth paths. They are neither finished nor accepted though, so we need something to do in the meantime.

The highbitdepth support seems not to be implemented yet ; but this looks promising!

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>