Make last_bipred_frame a reference frame
The frame marked as last_bipred at the encoder side is the frame right
before the source frame of BWDREF_FRAME, which was not used as a
reference. This CL is to add this type of frames to the reference frame
buffer and use them as reference for the coding of future frames. This
is a non-normative change.

Change-Id: If3dbce0959bb468e95d01a1feca1dac6bfb91d36