Make last_bipred_frame a reference frame
The frame marked as last_bipred at the encoder side is the frame right
before the source frame of BWDREF_FRAME, which was not used as a
reference. This CL is to add this type of frames to the reference frame
buffer and use them as reference for the coding of future frames. This
is a non-normative change.

A tiny gain is observed (lowres BDRate -0.016% in avg psnr). The main
target of this CL is to make it possible to have more flexible gf group
structure. For instance, LAST_BIPRED_UPDATE can be replaced by
LAST_UPDATE followed by an extra alt ref frame.

Change-Id: If3dbce0959bb468e95d01a1feca1dac6bfb91d36