ncobmc_adapt_weight: add rd functions to select ncobmc modes
These functions will be called at the final round of
encoding each superblock, wherein a ncobmc_adapr_weight
prediction will be built to check if it can provide
better results.

This commit contains two approaches to build the prediction.
One of them divides a squared block into four quadrants and
select an ncobmc mode for each quadrant. This method increases
the number of interpolation kernels we can used, and is used for
training purpose.

Change-Id: I117773a8133485362eebad5f3f55443cb5dc5136