Fix CMake shared library build.
The post relocation destination of some non-simd sources
that require specific simd flavors was incorrect. The
relocated sources touch internal APIs, and the incorrect
destination was an unguarded area where sources are added
unconditionally to targets with the assumption that all
APIs are accessible.

Sources moved to section guarded by BUILD_SHARED_LIBS, which
acts as the border between public APIs always available for
testing and those that require a static lib that exposes
private symbols.

Change-Id: I5d9ad513a0fe1b9dc57cb11cca53e8d00eb9b1a8