Update patch set 2
Patch Set 2:

> Do you need to set warp type to default for those ref frames, at
 > both gm search phase and bitstream packing phase, to make sure
 > about always using real zero mv?

It is already done at the gm search phase,  but I did forget it on the decoder side. I just added it and put in an assert when writing the parameters just to make sure it is always identity when the frames are different sizes.

Patch-set: 2
Reviewer: Gerrit User 5186 <5186@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>