Update patch set 1
Patch Set 1:

> 
> > This is a very good idea, and an argument for keeping a runtime flag to switch between the two paths.
> 
> Thanks, here it is: https://aomedia-review.googlesource.com/c/14960/
> 
> Please let me know what you think.
> James: if you're reading this, I'd be very interested to hear your opinion about this.
> 

I lost track of this. The tool might be helpful, but if it isn't automatically run what's catching the mismatch in the first place? As said this config flag seems strange and possibly incomplete in its usage in the library. It might be good to start a thread on the mailing list about what the intent is here.
This change could still be viewed as valid.

Patch-set: 1