Update patch set 35
Patch Set 35:

> Patch Set 35:
> 
> Hi Nathan,
> 
> Will you please let us know how much encoder/decoder speed would improve with the "reduced complexity"?

Hi Yi, I've taken over the Daala TX patchsets for the time being.

The reduction in complexity is through replacing the butterfly-style decomposition with a lifting topology.  There are fewer (and narrower) multiplies to the lifting topology, and it can run at a deeper bitdepth because it does not result in any scaling.  eg, we can and intend to run low and high bitdepth with the same transforms at a fixed coefficient shift (probably Q4, as that's what the transforms were designed to use in Daala).

That said, because it is a lifting transform, the arithmetic dependency depth is deeper than a butterfly-style transform.  This doesn't impact software implementations much, but we'll need to analyze carefully the impact on hardware.  Unlike the butterflies, the multiplies are not as parallelizable.

Monty

Patch-set: 35