Update patch set 35
Patch Set 35:

Oh, I should also have said-- due to my own oversight, the commit message is partly incorrect.  The metric numbers are up-to-date, but the text should have been updated to read 'there is no net coding performance impact'.  We expect moving to Q4 to be a modest net coding gain.  As they stand, running with the current Q3 shift, the coding performance results are a wash.

Patch-set: 35