Update patch set 35
Patch Set 35:

> on speed performance. According to your reply, lifting topology has
 > less multiplications but parallelism is not the same as butterfly.
 > So I still don't know if encoder/decoder will become faster or

Software will definitely be faster. The extra parallelism only matters for hardware. To a certain extent it might improve re-ordering opportunities in software, but transforms generally saturate the ALUs, so there's not much to gain from that.

The thing that does matter in software is that, not only are there fewer multiplies, but the complete multiply/downshift/round is a single instruction (VQRDMULH in NEON, PMULHRSW in SSSE3) that does not expand the operand sizes (doubling the SIMD throughput compared to multiplies whose output is twice as wide as the inputs).

However, I'd expect the total cost of transforms to be around 10% of the decode time in a modern video codec, so I doubt they will break the bank one way or the other.

Patch-set: 35
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>