(WIP) lgt: transforms based on intra prediction
In LGT experiments, sharp image discontinuity on the left/top
boundary is detected. Based on this discontinuity, we choose
different LGTs to apply to each row/column of the block. Different
rows/columns may have different transforms, depending on where
the image discontinuity potentially occurs.

This transform is implemented as a new tx_type, LGT2D and it can
be signaled in the RD search. Currently, LGT is only applied to
intra blocks.

Change-Id: I3e905bb4407ab3345db6dcba457af6194f46dfef