Update patch set 1
Patch Set 1:

> I lost track of this. The tool might be helpful, but if it isn't automatically run what's catching the mismatch in the first place?

It's meant to be automatically run.
We just can't enable it right now, as it fails.
Once Angie finally fixes the LBD/HBD mismatch (she might use this as a helper), I'll enable this test.

> As said this config flag seems strange

As I suggested above, the intent is to allow a user, for reference purposes, to prevent aomdec/aomenc from switching on its own between two incompatible coding paths when dealing with files using different profiles.

This is needed when you want to compare your HW with aomdec ; and that your HW implementation doesn't replicate the coding path duplication, and uses a unique coding path to decode all profiles.


> and possibly incomplete in its usage in the library.

What makes you say this ; are you thinking of something in particular?

I have strong reasons to believe this flag works:
1) a big part of our internal validation rely on it. 

2) the way it's implemented makes it hard to be incomplete: aomdec always had some logic to decide which coding path to use (depending on the profile).

I'm assuming this logic works, and that it's not duplicated anywhere else in the code.

The config flag just forces the decision. So there's no risk of having half of the decoder using LBD and the other half using HBD.

3) should it be incomplete, there would be crashes: the way a frame buffer handle is converted to a pixel pointer (e.g CONVERT_TO_SHORT_PTR) makes it nearly impossible to call a highbitdepth function with a lowbitdepth frame buffer without crashing (mainly because of the pointer shifting).


> This change could still be viewed as valid.

As a reminder, then, the following wouldn't pass:
./configure --disable-lowbitdepth && make && ./test_libaom

(because the encoder and the decoder would be using incompatible coding paths)

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>