Update patch set 5
Patch Set 5:

Am I right in thinking that this patch recalculates the "downconverted" frame for each reference on each frame? (ie, it does downconversions for the source + 6 refs per frame with EXT_REFS enabled)

In principle, each frame only has two distinct versions - the original version and the encoded version (note also that, in the presence of resizing, these may have different sizes!). So we should only need to downconvert each frame at most twice.

Since we only use the source version once, I'd like to propose that frm->y_buffer_8bit should only be used for the reference frames, and it should only be calculated once, the first time a frame is used as a reference. The downconverted source frame can just be stored as a local variable and discarded at the end of the function.

Patch-set: 5
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>