Allow automatic generation of uv_txsize_lookup
This patch has a new script, gen_uv_txsize_lookup.py, which generates
a uv_txsize_lookup array for use in common_data.h. See the header of
the script for detailed information about how it works.

At the moment, the only flag it ends up checking for the complicated
"Which transforms for this blocksize?" question is CONFIG_CHROMA_2X2,
so we generate the slightly messy "#if ! P/#endif/#if P/#endif" where
"#if P/#else/#endif" would have been better. However, we'll need to
handle the general case with more than one flag if we want to remove
the current restriction that square blocks don't become less square,
so I thought it was better to leave the general code doing its thing.

The patch also replaces the handwritten definition of uv_txsize_lookup
in common_data.h with the output of the script (which is in the form
expected by clang-format-diff).

There isn't yet any integration with the build system. The command I
used to manually do the replacement depends on knowing the line
numbers either side of the array:

  (A=av1/common/common_data.h; \
   git show origin/master:$A | head -n1125 >$A; \
   ./tools/gen_uv_txsize_lookup.py >>$A; \
   git show origin/master:$A | tail -n +1782 >>$A)

Change-Id: I85953ce56026c771bda72fd18af904e1f8812b12