Change default wedge signs
Change the default wedge sign so that the top-left mask entry
is always >= 32, rather than so that the average is >= 32.

This is intended to help with interintra prediction - such blocks
always use the default wedge sign, and we expect that the intra
prediction is usually better towards the top and left of the block.
Thus we want the signs to be set so that we generally use the intra
prediction on those edges and the inter prediction on the bottom and
right edges.

This should not impact inter-inter wedges, since those always send
a literal bit indicating which wedge sign to use.

Change-Id: Idf0e7b734d2adb1a9b56ba9d7c60012cfdfed972