[CFL] Reorganize Reconstructed Pixel Buffering
Reworked how the storage flag is set to avoid duplication on the encoder
side. Reconstructed Luma pixels are stored in encode_superblock in the
loop that calls av1_encode_intra_block_plane and in the extra call to
txfm_rd_in_plane after the luma RDO, but prior to the chroma RDO.

This change does not alter the bitsteam.

Change-Id: Ifd8441363ea0733fea3d06129a025940abb2abc9