Refactor iteration over neighbours for OBMC in reconinter.c
This moves the logic for "iterate over neighbouring blocks above/left"
into a pair of helper functions, into which is passed a function
pointer for the body. When compiled with optimisation, everything gets
inlined, so this is pretty similar to the original but without the
duplication in the source code.

The logic is also slightly generalised: the old code checked whether a
block was shorter or narrower than 8 pixels by comparing a block size
with BLOCK_8X8. This doesn't work when you have a 4x16 or 16x4 block
because e.g. BLOCK_16X4 is not less than BLOCK_8X8. This
generalisation is (unsurprisingly) needed in order to to support 16x4
or 4x16 blocks.

Finally, this patch doesn't currently address the CONFIG_NCOBMC
functions that do prediction from right and bottom edges.

Change-Id: I9e5a116b012c18645604a7d98fb98be99697d363