Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Did you confirm there is no performance change?

I haven't started an AWCY run, but have encoded a couple of images and got equal results (bit exact). This shouldn't be a surprise this patch just expresses the answers to "which mi?" and "is this for obmc?" directly rather than encoding them in a pair of offsets.

Patch-set: 3
Reviewer: Gerrit User 7560 <7560@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>