Change sgr to compute fewer mean / var samples
Skips the mean and variance computation for every other pixel
horizontally in sgr computation when radius is > 1 i.e. window-size is 5x5.
The missing mean / variance values are interpolated from neighboring
samples. This saves computation of these stats for half the samples.

In addition the last stage smoothing is also simplified for 5x5.

Change-Id: I5a0f855e7674ae6d96c3a2f413cbbbd70ff12293