Change sgr to compute fewer mean / var samples
Adds a macro to: skip the mean and variance computation for every other
row in sgr computation when radius is 2.
The missing mean / variance rows are interpolated from neighboring
samples vertically. This saves computation of these stats for half the
samples.

In addition the last stage smoothing is also simplified for radius 2
filter from 3x3 to 3x1.

Change-Id: Idb0482f9c3269a4a3ccb22b0e371feea93ea4cf2