Change sgr to compute fewer mean / var samples
Adds a macro to: skip the mean and variance computation for every other
pixel horizontally in sgr computation when radius is 2.
The missing mean / variance values are interpolated from neighboring
samples. This saves computation of these stats for half the samples.

In addition the last stage smoothing is also simplified for radius 2
filter from 3x3 to 1x3.

Change-Id: I5a0f855e7674ae6d96c3a2f413cbbbd70ff12293