Update patch set 1
Patch Set 1:

Maxym added the support for allowing users to use config files to specify the encoder options. This is the first step change to support the basic infrastructure. A simple flag is added to support turn on/off global motion and warped motion support as an example. We'd like to get some feedback from you regarding the basic infrastructure. if you are ok with it, we will start to expend it and add more control flags.

Patch-set: 1
Reviewer: Gerrit User 5326 <5326@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>