Fix tile boundary calculation
Fix a rare case in which the tile boundary information was not
set up properly in the decoder when using LOOPFILTERING_ACROSS_TILES

The situation was:
* One frame uses loop filtering across tiles. Then its tile
  boundary information is not needed, so is not calculated.
* The next frame (in decode order) has the same size and the
  same tile layout, but doesn't use loop filtering across tiles.
* Now the tile boundary information *is* needed, but we weren't
  recalculating it. This resulted in the loop filter being
  applied across tile boundaries even though we signalled not to.

To fix this issue, we cache whether the tile boundary information
has been set up for this frame size and tile layout. This allows
us to guarantee that the information is available when needed,
while minimizing recalculation.

Change-Id: I1f3cbb0537535bf38faaed4c21c07142e747f962