Fix ASan errors in SSE4.1 selfguided filter tests
The selfguided filter code was sometimes fetching 8 bytes of data
when it only needed 4. This was normally fine, but lead to problems
in the selfguided filter test when compiling for x86-32, where we
accidentally read off the end of the input buffer.

Fix this by only reading the amount of data we actually need.

BUG=aomedia:700

Change-Id: I2448b7b0d9cb2f9292a092675a66da64c89f913c