Encode loop restoration coefficients per tile
This is a baby-step towards encoding the coefficients at the start of
superblocks at the top-left of loop restoration tiles. Note that this
patch causes us to reset "wiener_info" and "sgrproj_info" at each tile
boundary, which will cause a performance drop.

This is necessary because, in order for tiles to be processed in
parallel, we cannot delta-encode coefficients across tile boundaries
if the coefficients are signalled within tiles. We could probably do
better than the current patch by, say, delta-encoding against previous
frames.

Note that this patch doesn't yet fix up the corresponding cost
estimates search_sgrproj and search_wiener in pickrst.c

Change-Id: I5b8b91d63aaf49627cde40219c31c0ac776dfd38