Correct scaling of 8x32 and 32x8 lowbitdepth transforms
According to av1_get_tx_scale() and comments, the 8x32 and 32x8
forward transforms should return coefficients with a scaling of 4x
unity, but they return 8x unity.  The inverse transforms match, so the
error is subtle, but this would affect RDO decisions.

The fix in this patch simply follows the shift pattern in other functions.
This bug does not exist in the 4x16 and 16x4 transforms.

Change-Id: I1535024e2fd118458807ba32fc1c75735bab59a3