Update patch set 1
Patch Set 1:

Thanks for catching the issue.
We were just looking at it.
Looks like the main issue was the av1_get_tx_scale() function in common/idct.c. That function should have returned 1 for 32x8/8x32. But we forgot to change that when we implemented the 4:1 transforms. Your patch also fixes it but then we lose precision of the coefficients that we could otherwise accommodate in 16 bits.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>