Update patch set 1
Patch Set 1:

@Jingning:

I've spent some time digging into the test failure and wanted to ask: Is mbmi->tx_size meant to be used for anything when the block of interest uses the recursive transform partitioning? I think my patch has affected how the decoder assigns mbmi->tx_size for var-tx blocks, and I'm not sure what the intended behaviour is.


Details of the situation I found below - apologies for it being long-winded!

What's going wrong is: We're decoding an intra block, so we end up in read_selected_tx_size to select one tx size for the whole block. This looks at xd->left_mbmi, which uses the recursive transform partitioning and contains a mixture of transform sizes. But we don't look at the inter_tx_size array, instead we look at the tx_size field.

On the decoder side, xd->left_mbmi->tx_size is 8x8, while on the encoder side it is 16x16. This leads to a mismatch in contexts and then in the decoded symbols. So the immediate cause is that mbmi->tx_size is set differently in the encoder and decoder.

But it seems weird to even look at this - wouldn't it be better to look at the transform size directly left of us? Or, since there might be multiple such transform blocks, maybe the smallest/largest of them?

Patch-set: 1
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>